App Exception Handling

+++ src/main/java/com/Product/caresale/service/CarService.java
@@ -1,5 +1,6 @@
 package com.Product.caresale.service;
 
+import com.Product.caresale.exception.ResourceNotFoundException;
 import com.Product.caresale.model.Car;
 import com.Product.caresale.repository.CarRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +11,7 @@ import java.util.List;
 @Service
 public class CarService {
     @Autowired
-    CarRepository carRepository;
+    CarRepository carRepository;  // create auto table to postgre sql
 
     public List<Car> getAllCars() {
         return  carRepository.findAll();
@@ -18,7 +19,17 @@ public class CarService {
     public Car saveCar(Car car) {
         return carRepository.save(car);
     }
+    public Car getCarById(Long id) {
+        return carRepository.findById(id)
+                .orElseThrow (() ->
+                   new ResourceNotFoundException(
+                           "Car with id " + id + " not found"));
+    }
     public void deleteCar(Long id) {
-        carRepository.deleteById(id);
+       if(!carRepository.existsById(id)) {
+           throw new ResourceNotFoundException(
+                   "Car with id " + id + "not found!");
+       }
+       carRepository.deleteById(id);
     }
 }
